:truck: Move creation of the babel-preset options to own module
